+++ sportorg/settings.py
@@ -38,16 +38,21 @@ class Settings:
     live_gzip_enabled: bool = True
 
     telegram_token: str = ""
+    teamwork_host: str = "localhost"
+    teamwork_port: int = 50010
+    teamwork_type_connection: str = "client"
+    teamwork_encryption_enabled: bool = False
+    teamwork_autorun: bool = False
+    teamwork_encryption_key: str = ""
+    teamwork_check_race_id: bool = False
 
-    source_names_path: str = config.base_dir("configs", "names.txt")
-    source_middle_names_path: str = config.base_dir("configs", "middle_names.txt")
-    source_regions_path: str = config.base_dir("configs", "regions.txt")
-    source_status_comments_path: str = config.base_dir("configs", "status_comments.txt")
-    source_status_default_comments_path: str = config.base_dir(
-        "configs", "status_default.txt"
-    )
-    source_ranking_score_path: str = config.base_dir("configs", "ranking.txt")
-    source_ranking_ardf_score_path: str = config.base_dir("configs", "ranking_ardf.txt")
+    source_names_path: str = config.configs_dir("names.txt")
+    source_middle_names_path: str = config.configs_dir("middle_names.txt")
+    source_regions_path: str = config.configs_dir("regions.txt")
+    source_status_comments_path: str = config.configs_dir("status_comments.txt")
+    source_status_default_comments_path: str = config.configs_dir("status_default.txt")
+    source_ranking_score_path: str = config.configs_dir("ranking.txt")
+    source_ranking_ardf_score_path: str = config.configs_dir("ranking_ardf.txt")
     source_rent_cards_path: str = config.data_dir("rent_cards.txt")